Fixed Bug -> Incorrect Workout History
Problem ->
 In the workout screen, when viewing the workout history for multiple movements similtaneously, all movements displayed the data from the most recently clicked movement.
This was caused by storing the result of the 'getMovementHistory' method in  the viewmodels state object as a 'UserDataHistoryState' sealed interface instance. This meant only one movement's history could be stored at any given time. With multiple composables inheriting the state object and accessing the UserDataHistoryState, this caused undesirable effects.

Solution -> 
The method has now been modified to return a Deferred<UserDataHistoryState>. The modfied method, now named 'getMovementHistoryAsync', is called from a LaunchedEffect block inside the 'AllMovementHistory' composable, the key parameter is set to the local variable 'numberOfSessions'. The LaunchedEffect block calls the method and assigns its return value to a 'remembered mutableState'. The 'more' button causes the method to be recalled by updating the 'numberOfSessions' variable.


Signed-off-by: Hamish Young <53566395+hyng3000@users.noreply.github.com>